# web/api/openapi.yaml
openapi: "3.0.0"
info:
  title: Seshat API
  description: 离线优先的 Bangumi 数据管理与浏览工具。所有数据从 Bangumi API 拉取并缓存本地。本地接口前缀 /api/v0，结构与官方 /v0 一致。backend_url 置空则前端请求本地 /api/v0，设为官方地址则直连官方 /v0。
  version: "0.1.0"
servers:
  - url: http://127.0.0.1:4000
    description: 本地服务器
...
paths:
...
  /api/v0/episodes:
    get:
      tags: [动画]
      summary: 列出动画关联的剧集列表
      parameters:
        - { name: subject_id, in: query, required: true, schema: { type: integer }, example: 51 }
        - { name: type, in: query, schema: { type: integer, description: "0=本篇 1=SP 2=OP 3=ED 4=预告 5=MAD 6=其他" }, example: 0 }
        - { name: limit, in: query, schema: { type: integer, maximum: 100, default: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: 剧集数组
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Episode"
...
  /api/v0/tags/{name}/subjects:
    get:
      tags: [标签]
      summary: 列出标签下的动画
      parameters: [{ name: name, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: 动画列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubjectListItem"

  # ── 图片 ──
  /api/v0/subjects/{id}/image:
    get:
      tags: [图片]
      summary: 动画图片
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: type, in: query, schema: { type: string, enum: [large, medium, small, grid, common], default: grid } }
      responses:
        "200": { description: 图片文件 }
        "404": { description: 未缓存，返回占位图 }
  /api/v0/characters/{id}/image:
    get:
      tags: [图片]
      summary: 角色图片
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: type, in: query, schema: { type: string, enum: [large, medium, small, grid], default: grid } }
      responses:
        "200": { description: 图片文件 }
        "404": { description: 未缓存，返回占位图 }
  /api/v0/persons/{id}/image:
    get:
      tags: [图片]
      summary: 人物图片
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: type, in: query, schema: { type: string, enum: [large, medium, small, grid], default: grid } }
      responses:
        "200": { description: 图片文件 }
        "404": { description: 未缓存，返回占位图 }

  # ── ELO ──
...
  /api/v0/elo/compare:
    post:
      tags: [ELO]
      summary: 提交比较结果
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                winner: { type: integer, example: 51 }
                loser: { type: integer, example: 288 }
      responses:
        "200": { description: OK }
...
  /api/v0/search/subjects:
    post:
      tags: [搜索]
      summary: 条目搜索
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                keyword: { type: string, example: "CLANNAD" }
                filter: { type: object, example: { type: [2] } }
      responses:
        "200":
          description: 搜索结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { type: object }
  /api/v0/search/characters:
    post:
      tags: [搜索]
      summary: 角色搜索
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                keyword: { type: string, example: "古河渚" }
      responses:
        "200":
          description: 搜索结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { type: object }
  /api/v0/search/persons:
    post:
      tags: [搜索]
      summary: 人物搜索
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                keyword: { type: string, example: "麻枝准" }
      responses:
        "200":
          description: 搜索结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { type: object }
  /api/v0/search/tags:
    post:
      tags: [搜索]
      summary: 标签搜索
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                keyword: { type: string, example: "治愈" }
      responses:
        "200":
          description: 搜索结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { type: object }

  # ── 拉取 ──
  /api/v0/fetch/subject:
    post:
      tags: [拉取]
      summary: 单个或批量拉取动画
      description: |
        从 Bangumi API 拉取动画的全部数据并缓存到本地。
        支持两种方式：
        - `{"id": 51}` 拉取单个
        - `{"ids": [51, 288, 9717]}` 批量拉取
        自动记录到 `_seshat.json`，返回 `task_id` 用于订阅实时进度。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: integer, example: 51, description: 单个动画 ID }
                ids: { type: array, items: { type: integer }, example: [51, 288, 9717], description: 多个动画 ID }
            examples:
              single:
                value: { id: 51 }
              batch:
                value: { ids: [51, 288, 9717] }
      responses:
        "200":
          description: 拉取已开始
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "fetching" }
                  count: { type: integer, example: 2 }
                  task_id: { type: string, example: "ae8687f0", description: "用于 GET /api/v0/task/{task_id}" }
...
  /api/v0/task/{task_id}:
    get:
      tags: [拉取]
      summary: 获取任务进度｜SSE
      description: 实时推送任务进度。同时只允许一个任务运行，新任务在已有任务完成前返回 409。
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
          example: "ae8687f0"
      responses:
        "200":
          description: text/event-stream 流式进度，首条消息含 task 类型和 label 描述
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  data: {"step":"start","done":0,"total":326,"task":"fetch_all","label":"刷新全部"}
                  data: {"step":"subjects","done":15,"total":326,"speed":"3.2/s","status":""}
                  data: {"step":"complete","done":326,"total":326}
        "409":
          description: 已有任务在运行
...
  /api/v0/fetch/tracker:
    post:
      tags: [拉取]
      summary: 按名称刷新指定 Tracker
      description: 传入 Tracker 名称数组，刷新其中的全部 subject。返回 task_id。
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                names: { type: array, items: { type: string }, example: ["seshat", "user"] }
      responses:
        "200":
          description: 后台执行
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  names: { type: array, items: { type: string } }
                  task_id: { type: string, example: "ae8687f0" }
  /api/v0/tracker/create:
    post:
      tags: [Tracker]
      summary: 创建空 Tracker
      description: 创建一个指定名称的空 `.toml` Tracker 文件，用户手动填入 ID。
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, example: "kyoani" }
      responses:
        "200":
          description: 创建成功
        "409":
          description: Tracker 已存在

  # ── 用户 ──
  /api/v0/users/{username}:
    get:
      tags: [用户]
      summary: 用户信息
      parameters:
        - { name: username, in: path, required: true, schema: { type: string }, example: "vanadiry" }
      responses:
        "200":
          description: 用户数据（不含 avatar）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /api/v0/users/{username}/avatar:
    get:
      tags: [用户]
      summary: 用户头像
      parameters:
        - { name: username, in: path, required: true, schema: { type: string }, example: "vanadiry" }
        - { name: type, in: query, schema: { type: string, enum: [large, medium, small], default: large } }
      responses:
        "200":
          description: 头像图片
          content:
            image/*:
              schema:
                type: string
                format: binary
...
components:
...
  schemas:
    SubjectListItem:
      type: object
      properties:
        id: { type: integer, example: 51 }
        name: { type: string, example: "CLANNAD -クラナド-" }
        name_cn: { type: string, example: "CLANNAD" }
        score: { type: number, example: 8.4 }
        platform: { type: string, example: "TV" }
        date: { type: string, example: "2007-10-04" }
        image: { type: string, example: "/api/v0/subjects/51/image?type=grid" }
...
    Episode:
      type: object
      properties:
        id: { type: integer, example: 1044433 }
        type: { type: integer, description: "0=本篇 1=SP 2=OP 3=ED 4=预告 5=MAD 6=其他", example: 0 }
        sort: { type: number, example: 1 }
        ep: { type: number, example: 1 }
        name: { type: string, example: "みんなでなら魔法少女になれる気がしたの" }
        name_cn: { type: string, example: "感觉和大家一起的话 就能够成为魔法少女" }
        duration: { type: string, example: "00:23:40" }
        airdate: { type: string, example: "2021-07-31" }
        desc: { type: string }
...
    User:
      type: object
      properties:
        id: { type: integer, example: 485742 }
        username: { type: string, example: "vanadiry" }
        nickname: { type: string }
        url: { type: string }
        sign: { type: string }
        user_group: { type: integer }
...